+++ service/src/main/kotlin/io/provenance/explorer/service/ExplorerService.kt
@@ -7,12 +7,13 @@ import cosmos.staking.v1beta1.Staking
 import cosmos.tx.v1beta1.ServiceOuterClass
 import io.provenance.explorer.config.ExplorerProperties
 import io.provenance.explorer.domain.core.logger
+import io.provenance.explorer.domain.entities.TransactionCacheRecord
 import io.provenance.explorer.domain.extensions.formattedString
 import io.provenance.explorer.domain.extensions.getStatusString
 import io.provenance.explorer.domain.extensions.height
 import io.provenance.explorer.domain.extensions.sendMsg
-import io.provenance.explorer.domain.extensions.signatureKey
 import io.provenance.explorer.domain.extensions.toOffset
+import io.provenance.explorer.domain.extensions.toSigObj
 import io.provenance.explorer.domain.extensions.toValue
 import io.provenance.explorer.domain.extensions.translateAddress
 import io.provenance.explorer.domain.extensions.translateByteArray
@@ -22,10 +23,12 @@ import io.provenance.explorer.domain.models.explorer.BlockDetail
 import io.provenance.explorer.domain.models.explorer.PagedResults
 import io.provenance.explorer.domain.models.explorer.RecentBlock
 import io.provenance.explorer.domain.models.explorer.RecentTx
+import io.provenance.explorer.domain.models.explorer.Signatures
 import io.provenance.explorer.domain.models.explorer.Spotlight
 import io.provenance.explorer.domain.models.explorer.TxDetails
 import io.provenance.explorer.domain.models.explorer.ValidatorSummary
-import io.provenance.explorer.grpc.toKeyValue
+import io.provenance.explorer.grpc.toMultiSig
+import io.provenance.explorer.grpc.toSingleSigKeyValue
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.async
 import kotlinx.coroutines.runBlocking
@@ -124,13 +127,13 @@ class ExplorerService(
     }
 
     private fun hydrateValidators(validators: List<Query.Validator>, stakingValidators: List<Staking.Validator>) = let {
-        val stakingPubKeys = stakingValidators.map { it.consensusPubkey.toKeyValue() }
+        val stakingPubKeys = stakingValidators.map { it.consensusPubkey.toSingleSigKeyValue() }
         val signingInfos = validatorService.getSigningInfos()
         val height = signingInfos.first().indexOffset
         val totalVotingPower = validators.sumBy { it.votingPower.toInt() }
-        validators.filter { stakingPubKeys.contains(it.pubKey.toKeyValue()) }
+        validators.filter { stakingPubKeys.contains(it.pubKey.toSingleSigKeyValue()) }
             .map { validator ->
-                val stakingValidator = stakingValidators.find { it.consensusPubkey.toKeyValue() == validator.pubKey.toKeyValue() }
+                val stakingValidator = stakingValidators.find { it.consensusPubkey.toSingleSigKeyValue() == validator.pubKey.toSingleSigKeyValue() }
                 val signingInfo = signingInfos.find { it.address == validator.address }
                 hydrateValidator(validator, stakingValidator!!, signingInfo!!, height.toInt(), totalVotingPower)
             }
@@ -177,7 +180,7 @@ class ExplorerService(
                 denomination = data.tx.tx.authInfo.fee.amountList[0].denom,
                 type = data.type,
                 blockHeight = data.tx.txResponse.height.toInt(),
-                signer = data.signer,
+                signers = TransactionCacheRecord.findSigsByHash(data.tx.txResponse.txhash).toSigObj(),
                 status = if (data.tx.txResponse.code > 0) "failed" else "success",
                 errorCode = data.tx.txResponse.code,
                 codespace = data.tx.txResponse.codespace)
@@ -193,28 +196,25 @@ class ExplorerService(
 
     fun getValidator(address: String) = validatorService.getValidator(address)
 
-    private fun hydrateTxDetails(tx: ServiceOuterClass.GetTxResponse) = let {
-        val signer = tx.tx.authInfo.signerInfosList.signatureKey()
-        TxDetails(
-            height = tx.txResponse.height.toInt(),
-            gasUsed = tx.txResponse.gasUsed.toInt(),
-            gasWanted = tx.txResponse.gasWanted.toInt(),
-            gasLimit = tx.tx.authInfo.fee.gasLimit.toInt(),
-            gasPrice = props.minGasPrice(),
-            time = blockService.getBlock(tx.txResponse.height.toInt())!!.block.header.time.formattedString(),
-            status = if (tx.txResponse.code > 0) "failed" else "success",
-            errorCode = tx.txResponse.code,
-            codespace = tx.txResponse.codespace,
-            fee = tx.tx.authInfo.fee.amountList[0].amount.toBigDecimal(),
-            feeDenomination = tx.tx.authInfo.fee.amountList[0].denom,
-            signer = signer?.toKeyValue(),
-            memo = tx.tx.body.memo,
-            txType = tx.txResponse.type()!!,
-            from = if (tx.txResponse.type() == "send") tx.sendMsg().fromAddress else "",
-            amount = if (tx.txResponse.type() == "send") tx.sendMsg().amountList[0].amount.toInt() else 0,
-            denomination = if (tx.txResponse.type() == "send") tx.sendMsg().amountList[0].denom else "",
-            to = if (tx.txResponse.type() == "send") tx.sendMsg().toAddress else "")
-    }
+    private fun hydrateTxDetails(tx: ServiceOuterClass.GetTxResponse) = TxDetails(
+        height = tx.txResponse.height.toInt(),
+        gasUsed = tx.txResponse.gasUsed.toInt(),
+        gasWanted = tx.txResponse.gasWanted.toInt(),
+        gasLimit = tx.tx.authInfo.fee.gasLimit.toInt(),
+        gasPrice = props.minGasPrice(),
+        time = blockService.getBlock(tx.txResponse.height.toInt())!!.block.header.time.formattedString(),
+        status = if (tx.txResponse.code > 0) "failed" else "success",
+        errorCode = tx.txResponse.code,
+        codespace = tx.txResponse.codespace,
+        fee = tx.tx.authInfo.fee.amountList[0].amount.toBigDecimal(),
+        feeDenomination = tx.tx.authInfo.fee.amountList[0].denom,
+        signers = TransactionCacheRecord.findSigsByHash(tx.txResponse.txhash).toSigObj(),
+        memo = tx.tx.body.memo,
+        txType = tx.txResponse.type()!!,
+        from = if (tx.txResponse.type() == "send") tx.sendMsg().fromAddress else "",
+        amount = if (tx.txResponse.type() == "send") tx.sendMsg().amountList[0].amount.toInt() else 0,
+        denomination = if (tx.txResponse.type() == "send") tx.sendMsg().amountList[0].denom else "",
+        to = if (tx.txResponse.type() == "send") tx.sendMsg().toAddress else "")
 
     fun getTransactionHistory(fromDate: DateTime, toDate: DateTime, granularity: String) =
         blockService.getTransactionCountsForDates(